Move alias logic to abstract class
